slack/requirements.txt
fastapi
python-dotenv
uvicorn
slack_bolt
